transacoes_bancarias/requirements.txt
Django == 3.2.20
djangorestframework == 3.13.1
psycopg2-binary == 2.9.6
requests ==  2.31.0
django-filter == 2.4.0
